+++ packages/amplify_datastore_plugin_interface/lib/src/publicTypes.dart
@@ -1,0 +1,6 @@
+export 'types/temporal/temporal_date.dart';
+export 'types/temporal/temporal_time.dart';
+export 'types/temporal/temporal_datetime.dart';
+export 'types/temporal/temporal_timestamp.dart';
+
+export 'types/exception/DataStoreException.dart';