+++ packages/amplify_api_plugin_interface/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.2-dev.2 (2021-02-09)
+
+### Features
+
+- feat: Implement the GraphQL Subscribe API (#338)
+
+### Chores
+
+- chore: Updated amplify-android version 1.6.10 (#332)
+- chore: Refactor error handling to use the new ApiException type (#339)
+
 ## 0.0.2-dev.1 (2021-01-25)
 
 ### Chores