+++ packages/amplify_core/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.2-dev.2 (2021-02-09)
+
+### Chores
+
+- chore: Introduce new AmplifyException type and Error utilities (#314)
+
+### Bug Fixes
+
+- fix: MissingPluginException when android app restarts (#345)
+
 ## 0.0.2-dev.1 (2021-01-25)
 
 ### Chores